[Sahil Ganeshwani] Add Lombok Library to auto

+++ src/main/java/com/bridgelabz/employeepayroll/dto/EmployeeDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class EmployeeDTO {
     private String name;
-    private double salary;
+    private long salary;
 }
